+++ objc2/src/lib.rs
@@ -168,6 +168,7 @@
     feature = "unstable-autoreleasesafe",
     feature(negative_impls, auto_traits)
 )]
+#![cfg_attr(feature = "unstable-c-unwind", feature(c_unwind))]
 #![warn(elided_lifetimes_in_paths)]
 #![warn(missing_docs)]
 #![deny(non_ascii_idents)]