+++ objc-sys/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Added `unstable-c-unwind` feature.
+
 
 ## 0.2.0-beta.0 - 2022-06-13